Form URLs with pathlib
Forming URLs with an f-string means that baseurl shouldn't end in a
trailing slash.  If it does then the URL will have consecutive slashes.
That's fine in the typical scenario where the URI maps to a path on disk
and the operating system ignores repeated path separators.  However, in
some cases this won't work, such as a repo hosted in S3.

Pathlib provides a more resilient way to form URLs.  There are various
path classes to choose from.  Since we won't be doing any filesystem I/O
and we want to use a slash (not a backslash) everywhere, PurePosixPath
makes the most sense.