Parse URLs before manipulating with pathlib
In 6876975 we switched to using `PurePosixPath` to form URLs.
Unfortunately that broke `load()`.  `urlopen` doesn't accept
`PurePosixPath` objects, and on top of that `PurePosixPath` concatenates
the consecutive slashes in the scheme.  `PurePosixPath` is still a good
solution for just the path part of a URL, as long as we parse that out
from the rest of the URL.  Fortunately we can use `urlparse` for this.

fixes #5