Add mirrorlist support
This patch adds mirrorlist support by introducing the
NotRepoException that is thrown when the HTTP call to
get the repomd.xml file returns 404.

The logic in load() is still the same, it first tries
the direct supplied URL then falls back trying to do
a call to the URL and parsing a list of HTTP URLs like
this mirrorlist [1].

It then loops through those URLs until it successfully
can grab a repomd.xml file.

This doesn't change any API for the library and has
complete test coverage.

[1] http://mirrorlist.centos.org/?repo=os&arch=x86_64&release=7